+++ mlox/modules/gui.py
@@ -1,4 +1,5 @@
 
+from __future__ import print_function
 import locale
 import os
 import sys
@@ -7,19 +8,18 @@ import codecs
 import traceback
 import StringIO
 import logging
-
 import wx
 import wx.richtext as rt
 import webbrowser
-webbrowser.PROCESS_CREATION_DELAY = 0
-
 import modules.version as version
+from modules.resources import gif_file, translation_file
 from modules.loadOrder import loadorder
 
-#Resource files
-program_path = os.path.realpath(sys.path[0])
-translation_file = os.path.join(program_path,"mlox.msg")
-gif_file = os.path.join(program_path,"mlox.gif")
+webbrowser.PROCESS_CREATION_DELAY = 0
+gui_logger = logging.getLogger('mlox.gui')
+
+#Set this as a constant, so we can easily change it later
+MLOX_IMG = wx.Image(gif_file, wx.BITMAP_TYPE_GIF)
 
 clip_file = "mlox_clipboard.out"
 debug_output = "mlox_debug.out"
@@ -162,8 +162,7 @@ class mlox_gui():
         self.frame.SetSizeHints(800,600)
         self.frame.SetBackgroundColour(wx.SystemSettings.GetColour(wx.SYS_COLOUR_3DFACE))
         # logo doubles as a "reload" button
-        img = wx.Image(gif_file, wx.BITMAP_TYPE_GIF).ConvertToBitmap()
-        self.logo = wx.BitmapButton(self.frame, -1, img, (0,0), (img.GetWidth()+5, img.GetHeight()+5))
+        self.logo = wx.BitmapButton(self.frame, -1, MLOX_IMG.ConvertToBitmap(), (0,0), (MLOX_IMG.GetWidth()+5, MLOX_IMG.GetHeight()+5))
         self.logo.Bind(wx.EVT_BUTTON, self.on_reload)
         self.logo.SetToolTip(wx.ToolTip(_["Click to Reload"]))
         self.label_stats = wx.StaticText(self.frame, -1, _["Statistics"])
@@ -273,6 +272,19 @@ class mlox_gui():
         the_url = e.GetString()
         webbrowser.open(the_url)
 
+    #Display output messages on the gui panels (and disable/enable btn_update if appropriate)
+    def display(self):
+        if self.can_update:
+            self.btn_update.Enable()
+        else:
+            self.btn_update.Disable()
+        display_colored_text(self.Stats.getvalue(),self.txt_stats)
+        display_colored_text(self.Msg.getvalue(),self.txt_msg)
+        display_colored_text(self.New.getvalue(),self.txt_new)
+        self.txt_cur.SetValue(self.Old.getvalue())
+        self.label_cur.SetLabel(self.lo.origin)
+        self.cur_vbox.Layout()
+
     def analyze_loadorder(self, fromfile):
         #Clear all the outputs (except Dbg)
         self.New.truncate(0)
@@ -280,7 +292,7 @@ class mlox_gui():
         self.Stats.truncate(0)
         self.Msg.truncate(0)
 
-        logging.info("Version: %s\t\t\t\t %s " % (version.full_version(), _["Hello!"]))
+        gui_logger.info("Version: %s\t\t\t\t %s " % (version.full_version(), _["Hello!"]))
         self.lo = loadorder()
         if fromfile != None:
             self.lo.read_from_file(fromfile)
@@ -290,7 +302,8 @@ class mlox_gui():
                 self.lo.get_data_files()
         progress = wx.ProgressDialog("Progress", "", 100, None,
                                          wx.PD_AUTO_HIDE|wx.PD_APP_MODAL|wx.PD_ELAPSED_TIME)
-        self.lo.update(self.Msg,progress)
+        #print(self.lo.update())
+        print(self.lo.update(progress), file=self.Msg)
         progress.Destroy()
         for p in self.lo.get_original_order():
             self.Old.write(p+'\n')
@@ -300,14 +313,7 @@ class mlox_gui():
             self.can_update = False
 
         #Go ahead and display everything
-        if not self.can_update:
-            self.btn_update.Disable()
-        display_colored_text(self.Stats.getvalue(),self.txt_stats)
-        display_colored_text(self.Msg.getvalue(),self.txt_msg)
-        display_colored_text(self.New.getvalue(),self.txt_new)
-        self.txt_cur.SetValue(self.Old.getvalue())
-        self.label_cur.SetLabel(self.lo.origin)
-        self.cur_vbox.Layout()
+        self.display()
 
     def start(self):
         self.frame.Show(True)
@@ -327,9 +333,9 @@ class mlox_gui():
         if not self.can_update:
             return
         self.lo.write_new_order()
-        logging.info("[LOAD ORDER UPDATED!]")
+        gui_logger.info("[LOAD ORDER UPDATED!]")
         self.can_update = False
-        self.btn_update.Disable()
+        self.display()
 
     def on_close(self, e):
         self.on_quit(e)
@@ -338,8 +344,8 @@ class mlox_gui():
         try:
             out = open(debug_output, 'w')
         except IOError:
-            logging.error("Unable to write to debug output file:  {0}".format(debug_output))
-        print >> out, self.Dbg.getvalue().encode("utf-8")
+            gui_logger.error("Unable to write to debug output file:  {0}".format(debug_output))
+        print(self.Dbg.getvalue().encode("utf-8"), file=out)
         out.close()
 
     def right_click_handler(self, e):
@@ -368,7 +374,7 @@ class mlox_gui():
                     try:
                         out = open(clip_file, 'w')
                     except IOError:
-                        logging.error("Unable to open temporary clipboard file:  {0}".format(clip_file))
+                        gui_logger.error("Unable to open temporary clipboard file:  {0}".format(clip_file))
                         wx.TheClipboard.Close()
                         return
                     # sometimes some unicode muck can get in there, as when pasting from web pages.