+++ mlox/modules/update.py
@@ -1,9 +1,6 @@
-# Update mlox_base.txt from the internet
+# Update a file from the internet
+# with the option of auto extracting it
 
-UPDATE_FILE = 'mlox-data.7z'
-UPDATE_URL = 'https://sourceforge.net/projects/mlox/files/mlox/' + UPDATE_FILE
-
-import sys
 import os
 import urllib
 import subprocess
@@ -11,50 +8,53 @@ import logging
 
 update_logger = logging.getLogger('mlox.update')
 
-
-#Download UPDATE_URL to the program's main directory, then extract its contents there
-def update_mloxdata():
-    program_path = os.path.realpath(sys.path[0])
-    full_update_file = os.path.join(program_path,UPDATE_FILE)
-
-    update_logger.info('Checking for database update...')
+#Check if a file has changed (size) compared to local version
+def isNewer(local_file,url):
+    if not os.path.isfile(local_file):
+        return True
     try:
-        d = urllib.urlopen(UPDATE_URL)
+        connection = urllib.urlopen(url)
     except:
-        update_logger.warning('Unable to connect to {0}, skipping mlox data update.'.format(UPDATE_URL))
-        return
+        update_logger.warning('Unable to connect to {0}, skipping update.'.format(url))
+        return False
 
-    #Compare file sizes to see if an update is needed or not
-    if os.path.isfile(full_update_file):
-        fsize = os.stat(full_update_file).st_size
-        update_logger.debug('Current size: {0}'.format(fsize))
-        dsize = d.info()['Content-Length']
-        update_logger.debug('Downloadable size: {0}'.format(dsize))
-        update = int(dsize) != int(fsize)
-    else:
-        update = 1
+    local_size = os.stat(local_file).st_size
+    update_logger.debug('Current size: {0}'.format(local_size))
+    url_size = connection.info()['Content-Length']
+    update_logger.debug('Downloadable size: {0}'.format(url_size))
 
-    if update:
-        update_logger.info('Updating {0}'.format(full_update_file))
+    return int(url_size) != int(local_size)
+
+#Extract the contents of a file to a directory (using 7za)
+def extract(file_path,directory):
+    cmd = ['7za', 'e', '-aoa', '-o{0}'.format(directory), file_path]
+    update_logger.debug("Extracting via command %s",cmd)
+    try:
+        devnull = open(os.devnull, 'w')
+        subprocess.check_call(cmd, stdout=devnull)
+    except Exception as e:
+        update_logger.error('Error while extracting from {0}'.format(file_path))
+        #update_logger.error('Exception {0} while trying to execute command:  {0}'.format(str(e),cmd))
+        return False
+    return True
+
+#Check if a compressed file needs updating.
+#If it does, download it, and extract it to a directory
+def update_compressed_file(file_path,url,directory):
+    update_needed = isNewer(file_path,url)
+    if update_needed:
+        update_logger.info('Updating {0}'.format(file_path))
         try:
-            d = urllib.urlretrieve(UPDATE_URL,full_update_file)
-        except:
-            update_logger.error('Unable to download {0}, skipping mlox data update'.format(UPDATE_URL))
-            return
-
-        update_logger.info('Downloaded {0}'.format(full_update_file))
-
-        #Extract the file
-        cmd = ['7za', 'e', '-aoa', '-o{0}'.format(program_path), full_update_file]
-        try:
-            devnull = open(os.devnull, 'w')
-            subprocess.check_call(cmd, stdout=devnull)
+            urllib.urlretrieve(url,file_path)
         except Exception as e:
-            update_logger.error('Error while extracting from {0}'.format(full_update_file))
-            #update_logger.error('Exception {0} while trying to execute command:  {0}'.format(str(e),cmd))
-            return
-        update_logger.info('mlox_base.txt updated from {0}'.format(full_update_file))
+            update_logger.error('Unable to download {0}, skipping update.'.format(url))
+            update_logger.debug('Error: {0}'.format(e))
+            return False
 
+        update_logger.info('Downloaded {0}'.format(file_path))
+        if not extract(file_path,directory):
+            return False
+        return True
     else:
-        update_logger.info('No update necessary for file {0}'.format(full_update_file))
-    return
+        update_logger.info('No update necessary for file {0}'.format(file_path))
+        return False